add interactivity with scss and js
The biggest update so far! I incorporated Sass into the site which allowed me to generate the drag-and-drop images in a random position throughout the page. Using javascript I was able to generate this drag-and-drop functionality by manipulating the DOM. I learaned that I love Sass and will continue to use it to style any future projects.